Do not unnecessarily cast format parameters
Types like strong enumerations and std::uint8_t are gracefully handled
by the formatter.

+++ fly/coders/huffman/huffman_decoder.cpp
@@ -46,9 +46,7 @@ bool HuffmanDecoder::decode_binary(BitStreamReader &encoded, std::ostream &decod
 
         if (!decode_codes(encoded, max_code_length))
         {
-            LOGW(
-                "Error decoding codes from stream (maximum code length = {})",
-                static_cast<std::uint32_t>(m_max_code_length));
+            LOGW("Error decoding codes from stream (maximum code length = {})", m_max_code_length);
             return false;
         }
         else if (!decode_symbols(encoded, max_code_length, chunk_size, decoded))
@@ -82,7 +80,7 @@ bool HuffmanDecoder::decode_header(BitStreamReader &encoded, std::uint32_t &chun
             return decode_header_version1(encoded, chunk_size);
 
         default:
-            LOGW("Decoded invalid Huffman version {}", static_cast<std::uint32_t>(huffman_version));
+            LOGW("Decoded invalid Huffman version {}", huffman_version);
             break;
     }
 
@@ -102,7 +100,7 @@ bool HuffmanDecoder::decode_header_version1(BitStreamReader &encoded, std::uint3
     }
     else if (encoded_chunk_size_kb == 0)
     {
-        LOGW("Decoded invalid chunk size {}", static_cast<std::uint32_t>(encoded_chunk_size_kb));
+        LOGW("Decoded invalid chunk size {}", encoded_chunk_size_kb);
         return false;
     }
 
@@ -118,9 +116,7 @@ bool HuffmanDecoder::decode_header_version1(BitStreamReader &encoded, std::uint3
         (encoded_max_code_length == 0) ||
         (encoded_max_code_length >= std::numeric_limits<code_type>::digits))
     {
-        LOGW(
-            "Decoded invalid maximum code length {}",
-            static_cast<std::uint32_t>(encoded_max_code_length));
+        LOGW("Decoded invalid maximum code length {}", encoded_max_code_length);
         return false;
     }
 
@@ -145,9 +141,7 @@ bool HuffmanDecoder::decode_codes(BitStreamReader &encoded, length_type &max_cod
     }
     else if ((counts_size == 0) || (counts_size > (m_max_code_length + 1)))
     {
-        LOGW(
-            "Decoded invalid number of code length counts {}",
-            static_cast<std::uint32_t>(counts_size));
+        LOGW("Decoded invalid number of code length counts {}", counts_size);
         return false;
     }
 
@@ -176,9 +170,7 @@ bool HuffmanDecoder::decode_codes(BitStreamReader &encoded, length_type &max_cod
 
             if (!encoded.read_byte(symbol))
             {
-                LOGW(
-                    "Could not decode symbol of length {} bits",
-                    static_cast<std::uint32_t>(length));
+                LOGW("Could not decode symbol of length {} bits", length);
                 return false;
             }
